Simplify uploadObservation code by moving some logic into Obs model
