+++ src/components/Observations/UploadProgressBar.js
@@ -3,58 +3,31 @@
 import BottomSheet, { BottomSheetView } from "@gorhom/bottom-sheet";
 import { t } from "i18next";
 import type { Node } from "react";
-import React, {
-  useCallback, useEffect, useMemo, useRef, useState
-} from "react";
+import React, { useMemo, useRef } from "react";
 import {
   Button, ProgressBar, Text
 } from "react-native-paper";
 
-import Observation from "../../models/Observation";
-import uploadObservation from "../../providers/uploadHelpers/uploadObservation";
 import colors from "../../styles/colors";
 import { textStyles, viewStyles } from "../../styles/observations/uploadProgressBar";
+import useUploadObservations from "./hooks/useUploadObservations";
 
 type Props = {
   uploadStatus: Object
 }
 
 const UploadProgressBar = ( { uploadStatus }: Props ): Node => {
-  const { allObsToUpload, unuploadedObs, totalObsToUpload } = uploadStatus;
+  const { unuploadedObs } = uploadStatus;
   const numOfUnuploadedObs = unuploadedObs?.length;
-  const [cancelUpload, setCancelUpload] = useState( false );
-  const [currentUploadIndex, setCurrentUploadIndex] = useState( 0 );
-  const [status, setStatus] = useState( null );
-
-  const calculateProgress = ( ) => ( totalObsToUpload - numOfUnuploadedObs ) / totalObsToUpload;
-  const progressFraction = calculateProgress( );
-
-  useEffect( ( ) => {
-    const upload = async obs => {
-      const mappedObs = Observation.mapObservationForUpload( obs );
-      const uploadSuccess = await uploadObservation( mappedObs, obs );
-      if ( uploadSuccess !== "success" ) {
-        setStatus( uploadSuccess );
-        return;
-      }
-      if ( currentUploadIndex < allObsToUpload.length - 1 ) {
-        setCurrentUploadIndex( currentUploadIndex + 1 );
-      }
-    };
-
-    if ( !cancelUpload ) {
-      upload( allObsToUpload[currentUploadIndex] );
-    }
-  }, [cancelUpload, currentUploadIndex, allObsToUpload] );
+  const {
+    handleClosePress,
+    status,
+    progressFraction
+  } = useUploadObservations( uploadStatus );
 
   const sheetRef = useRef( null );
 
   const snapPoints = useMemo( () => ["25%"], [] );
-
-  const handleClosePress = useCallback( () => {
-    setCancelUpload( true );
-    setStatus( null );
-  }, [] );
 
   // eslint-disable-next-line react/jsx-no-useless-fragment
   const noHandle = ( ) => <></>;
