+++ src/providers/ObsEditProvider.js
@@ -113,8 +113,7 @@ const ObsEditProvider = ( { children }: Props ): Node => {
 
     const saveAndUploadObservation = async ( ) => {
       const localObs = await saveLocalObservation( currentObs );
-      const mappedObs = Observation.mapObservationForUpload( localObs );
-      uploadObservation( mappedObs, localObs );
+      uploadObservation( localObs );
       if ( localObs ) {
         setNextScreen( );
       }
